requirements.txt
mcp
pinecone-client
sentence-transformers
langchain-pinecone
langchain-huggingface
langchain-community
langchain-text-splitters
pypdf
docx2txt
python-pptx
